Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@
 int main() {
     float a, b, c;
     scanf("%f %f %f", &a, &b, &c);
-    int sum= a+b+c;
+    float sum= a+b+c;
     printf("Average: %.2f", sum/3);
     return 0;
 }
